Refactor attack result into orchestrator result
